@@ -1,9 +1,9 @@
 package action
 
 import (
+	"github.com/devttl/goctl-swagger/generate"
 	"github.com/urfave/cli/v2"
 	"github.com/zeromicro/go-zero/tools/goctl/plugin"
-	"github.com/zeromicro/goctl-swagger/generate"
 )
 
 func Generator(ctx *cli.Context) error {
@@ -20,5 +20,7 @@ func Generator(ctx *cli.Context) error {
 	basepath := ctx.String("basepath")
 	host := ctx.String("host")
 	schemes := ctx.String("schemes")
-	return generate.Do(fileName, host, basepath, schemes, p)
+	pack := ctx.String("pack")
+	response := ctx.String("response")
+	return generate.Do(fileName, host, basepath, schemes, pack, response, p)
 }
